Modify unit tests to cover new shardHandler
